Move theodo coordinates from "utils" to "constant"
