+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const positionTheodo: [number, number] = [48.882737, 2.322391];
